fix(config parse): make sure to use floats

+++ src/HarvestBalanceCalculator/WorkingPreference.py
@@ -3,7 +3,7 @@ import os
 
 class WorkingPreference:
     def __init__(self):
-        self.workdays = list(map(int, os.environ['WORK_DAYS_OF_THE_WEEK'].split()))
+        self.workdays = list(map(float, os.environ['WORK_DAYS_OF_THE_WEEK'].split()))
         self.hours_per_day = float(os.environ['HOURS_PER_WORK_DAY'])
 
     def is_a_working_day(self, date):
